Added finals and nonnull annotations, moved VanishHook to the api
