+++ api/src/main/java/at/helpch/chatchat/api/hook/Hook.java
@@ -1,4 +1,4 @@
-package at.helpch.chatchat.api;
+package at.helpch.chatchat.api.hook;
 
 import java.util.List;
 import java.util.Optional;
