+++ plugin/src/main/java/at/helpch/chatchat/hooks/dsrv/ChatChatDsrvHook.java
@@ -1,7 +1,7 @@
 package at.helpch.chatchat.hooks.dsrv;
 
 import at.helpch.chatchat.ChatChatPlugin;
-import at.helpch.chatchat.api.Hook;
+import at.helpch.chatchat.api.hook.Hook;
 import github.scarsz.discordsrv.DiscordSRV;
 import java.util.List;
 import java.util.Optional;
@@ -12,7 +12,7 @@ public final class ChatChatDsrvHook implements Hook {
 
     private final ChatChatPlugin plugin;
 
-    public ChatChatDsrvHook(final ChatChatPlugin plugin) {
+    public ChatChatDsrvHook(@NotNull final ChatChatPlugin plugin) {
         this.plugin = plugin;
     }
