+++ plugin/src/main/java/at/helpch/chatchat/hooks/towny/ChatChatTownyHook.java
@@ -1,7 +1,7 @@
 package at.helpch.chatchat.hooks.towny;
 
 import at.helpch.chatchat.ChatChatPlugin;
-import at.helpch.chatchat.api.Hook;
+import at.helpch.chatchat.api.hook.Hook;
 import java.util.List;
 import java.util.Optional;
 import org.jetbrains.annotations.NotNull;
@@ -10,7 +10,7 @@ public class ChatChatTownyHook implements Hook {
 
     private final ChatChatPlugin plugin;
 
-    public ChatChatTownyHook(ChatChatPlugin plugin) {
+    public ChatChatTownyHook(@NotNull final ChatChatPlugin plugin) {
         this.plugin = plugin;
     }
